Component: Add FlyingLegend, Refactor Planning-Legend
- Extract flying legend to component
- Standardize import/export of planning legend
