+++ src/components/MapLegend/IncidentLegendGrid.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import {
   CadastreLegendItemPoint,
   HeaderCol,
