+++ src/components/MapLegend/index.ts
@@ -1,16 +1,4 @@
-import ConstructionLegendImage from './assets/construction.png';
-import DividersLegendImage from './assets/dividers.png';
-import FootwaysLegendImage from './assets/footways.png';
-import GuardrailLegendImage from './assets/guardrail-legend.png';
-
 export * from './Legend';
 export * from './IncidentLegendGrid';
 export * from './CadastreLegendGrid';
 export * from './LegendStyles';
-
-export {
-  ConstructionLegendImage,
-  DividersLegendImage,
-  FootwaysLegendImage,
-  GuardrailLegendImage,
-};
