+++ src/components2/MapLegend/index.ts
@@ -1,0 +1,1 @@
+export * from './FlyingLegend';
